requirements.txt
attrs==21.4.0
aws-cdk.cloud-assembly-schema==1.143.0
aws-cdk.core==1.143.0
aws-cdk.cx-api==1.143.0
aws-cdk.region-info==1.143.0
cattrs==1.10.0
constructs==3.3.207
jsii==1.52.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
attrs==21.4.0
aws-cdk.assets==1.143.0
aws-cdk.aws-cloudwatch==1.143.0
aws-cdk.aws-ec2==1.143.0
aws-cdk.aws-events==1.143.0
aws-cdk.aws-iam==1.143.0
aws-cdk.aws-kms==1.143.0
aws-cdk.aws-logs==1.143.0
aws-cdk.aws-s3==1.143.0
aws-cdk.aws-s3-assets==1.143.0
aws-cdk.aws-ssm==1.143.0
aws-cdk.cloud-assembly-schema==1.143.0
aws-cdk.core==1.143.0
aws-cdk.cx-api==1.143.0
aws-cdk.region-info==1.143.0
cattrs==1.10.0
constructs==3.3.207
jsii==1.52.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
